+++ modules/config/default.js
@@ -38,6 +38,8 @@ export const settings = {
   canReorder: true,
   canRegroup: true,
   showNot: true,
+  canLeaveEmptyGroup: true,
+  shouldCreateEmptyGroup: false,
   canShortMongoQuery: true,
   groupActionsPosition: "topRight", // oneOf [topLeft, topCenter, topRight, bottomLeft, bottomCenter, bottomRight]
   setOpOnChangeField: ["keep", "default"], // 'default' (default if present), 'keep' (keep prev from last field), 'first', 'none'