+++ modules/utils/stuff.js
@@ -291,3 +291,15 @@ export function sleep(delay) {
     setTimeout(resolve, delay);
   });
 }
+
+export const isImmutable = (v) => {
+  return typeof v === "object" && v !== null && typeof v.toJS === "function";
+};
+
+export function applyToJS(v) {
+  return (isImmutable(v) ? v.toJS() : v);
+}
+
+export function toImmutableList(v) {
+  return (isImmutable(v) ? v : new Immutable.List(v));
+}